Remove actions
Remove actions.
